feat: remove duplicated code (PIN-10746)
