+++ src/components/dialogs/Dialog.tsx
@@ -40,10 +40,8 @@ import type {
   DialogArchiveVersionProps,
   DialogCancelVersionArchivingProps,
   DialogBlockArchivingRequestProps,
-  DialogDelegatorConfirmArchivingEServiceProps,
-  DialogDelegatorConfirmArchivingVersionProps,
-  DialogDelegatorRejectArchivingEServiceProps,
-  DialogDelegatorRejectArchivingVersionProps,
+  DialogDelegatorConfirmArchivingProps,
+  DialogDelegatorRejectArchivingProps,
 } from '@/types/dialog.types'
 import { DialogRejectAgreement } from './DialogRejectAgreement'
 import { ErrorBoundary } from '../shared/ErrorBoundary'
@@ -79,10 +77,8 @@ import { DialogReactivateArchivingDescriptor } from './DialogReactivateArchiving
 import { DialogArchiveVersion } from './DialogArchiveVersion'
 import { DialogCancelVersionArchiving } from './DialogCancelVersionArchiving'
 import { DialogBlockArchivingRequest } from './DialogBlockArchivingRequest'
-import DialogDelegatorConfirmArchivingEService from './DialogDelegatorConfirmArchivingEService'
-import DialogDelegatorConfirmArchivingVersion from './DialogDelegatorConfirmArchivingVersion'
-import DialogDelegatorRejectArchivingEService from './DialogDelegatorRejectArchivingEService'
-import DialogDelegatorRejectArchivingVersion from './DialogDelegatorRejectArchivingVersion'
+import DialogDelegatorConfirmArchiving from './DialogDelegatorConfirmArchiving'
+import DialogDelegatorRejectArchiving from './DialogDelegatorRejectArchiving'
 
 function match<T>(
   onBasic: (props: DialogBasicProps) => T,
@@ -120,10 +116,8 @@ function match<T>(
   onArchiveVersion: (props: DialogArchiveVersionProps) => T,
   onCancelVersionArchiving: (props: DialogCancelVersionArchivingProps) => T,
   onBlockArchivingRequest: (props: DialogBlockArchivingRequestProps) => T,
-  onDelegatorConfirmArchivingEService: (props: DialogDelegatorConfirmArchivingEServiceProps) => T,
-  onDelegatorConfirmArchivingVersion: (props: DialogDelegatorConfirmArchivingVersionProps) => T,
-  onDelegatorRejectArchivingEService: (props: DialogDelegatorRejectArchivingEServiceProps) => T,
-  onDelegatorRejectArchivingVersion: (props: DialogDelegatorRejectArchivingVersionProps) => T
+  onDelegatorConfirmArchiving: (props: DialogDelegatorConfirmArchivingProps) => T,
+  onDelegatorRejectArchiving: (props: DialogDelegatorRejectArchivingProps) => T
 ) {
   return (props: DialogProps) => {
     switch (props.type) {
@@ -197,14 +191,10 @@ function match<T>(
         return onCancelVersionArchiving(props)
       case 'blockArchivingRequest':
         return onBlockArchivingRequest(props)
-      case 'delegatorConfirmArchivingEService':
-        return onDelegatorConfirmArchivingEService(props)
-      case 'delegatorConfirmArchivingVersion':
-        return onDelegatorConfirmArchivingVersion(props)
-      case 'delegatorRejectArchivingEService':
-        return onDelegatorRejectArchivingEService(props)
-      case 'delegatorRejectArchivingVersion':
-        return onDelegatorRejectArchivingVersion(props)
+      case 'delegatorConfirmArchiving':
+        return onDelegatorConfirmArchiving(props)
+      case 'delegatorRejectArchiving':
+        return onDelegatorRejectArchiving(props)
     }
   }
 }
@@ -245,10 +235,8 @@ const _Dialog = match(
   (props) => <DialogArchiveVersion {...props} />,
   (props) => <DialogCancelVersionArchiving {...props} />,
   () => <DialogBlockArchivingRequest />,
-  (props) => <DialogDelegatorConfirmArchivingEService {...props} />,
-  (props) => <DialogDelegatorConfirmArchivingVersion {...props} />,
-  (props) => <DialogDelegatorRejectArchivingEService {...props} />,
-  (props) => <DialogDelegatorRejectArchivingVersion {...props} />
+  (props) => <DialogDelegatorConfirmArchiving {...props} />,
+  (props) => <DialogDelegatorRejectArchiving {...props} />
 )
 
 export const Dialog: React.FC = () => {
